BUG: input must be signed. ENH: Better documentation.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=12832
git-svn-id: http://svn.slicer.org/Slicer4/trunk@12832 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/CLI/OtsuThresholdSegmentation.cxx
@@ -49,9 +49,9 @@ int main( int argc, char * argv[] )
   PARSE_ARGS;
 
 // Image Types
-  typedef  unsigned long  InputPixelType;
-  typedef  unsigned long  InternalPixelType;
-  typedef  short  OutputPixelType;
+  typedef  long  InputPixelType;
+  typedef  long  InternalPixelType;
+  typedef  long  OutputPixelType;
 
   typedef itk::Image< InputPixelType,  3 >   InputImageType;
   typedef itk::Image< InternalPixelType, 3 >   InternalImageType;
